+++ app/Http/Controllers/ListingController.php
@@ -1,0 +1,98 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Listing;
+use Illuminate\Http\Request;
+
+class ListingController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function view()
+    {
+       return view('admin.listing');
+    }
+    public function index()
+    {
+        $listing = Listing::get();
+        return response()->json($listing);
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     * 
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Listing  $listing
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Listing $listing)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\Listing  $listing
+     * @return \Illuminate\Http\Response
+     */
+    public function edit(Listing $listing)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Listing  $listing
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, Listing $listing)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Listing  $listing
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+       
+        try {
+            $category = Listing::find($id);
+            $category->delete();
+            return response()->json("['status' => 'success']");
+        } catch (\Throwable $th) {
+            return response()->json("['status' => 'error']");
+        }
+    }
+}
