+++ resources/views/admin/layouts/main.blade.php
@@ -1,3 +1,4 @@
 @include('admin.layouts.header')
 @yield('main-section')
 @include('admin.layouts.footer')
+@yield('section-script')
