+++ resources/views/admin/listing.blade.php
@@ -24,618 +24,20 @@
                                         <table id="example3" class="display" style="min-width: 845px">
                                             <thead>
                                                 <tr>
-                                                    <th></th>
-                                                    <th>Name</th>
-                                                    <th>Department</th>
-                                                    <th>Gender</th>
-                                                    <th>Education</th>
-                                                    <th>Mobile</th>
-                                                    <th>Email</th>
-                                                    <th>Joining Date</th>
+                                                    <th>ID</th>
+                                                    <th>Title</th>
+                                                    <th>Type</th>
+                                                    <th>Category</th>
+                                                    <th>Address</th>
+                                                    <th>City</th>
+                                                    <th>Package</th>
+                                                    <th>Description</th>
+                                                    <th>Slug</th>
                                                     <th>Action</th>
                                                 </tr>
                                             </thead>
-                                            <tbody>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic1.jpg" alt=""></td>
-                                                    <td>Tiger Nixon</td>
-                                                    <td>Architect</td>
-                                                    <td>Male</td>
-                                                    <td>M.COM., P.H.D.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2011/04/25</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic2.jpg" alt=""></td>
-                                                    <td>Garrett Winters</td>
-                                                    <td>Accountant</td>
-                                                    <td>Female</td>
-                                                    <td>M.COM., P.H.D.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2011/07/25</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic3.jpg" alt=""></td>
-                                                    <td>Ashton Cox</td>
-                                                    <td>Junior Technical</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2009/01/12</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic4.jpg" alt=""></td>
-                                                    <td>Cedric Kelly</td>
-                                                    <td>Developer</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/03/29</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic5.jpg" alt=""></td>
-                                                    <td>Airi Satou</td>
-                                                    <td>Accountant</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2008/11/28</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic6.jpg" alt=""></td>
-                                                    <td>Brielle Williamson</td>
-                                                    <td>Specialist</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/12/02</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic7.jpg" alt=""></td>
-                                                    <td>Herrod Chandler</td>
-                                                    <td>Sales Assistant</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/08/06</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic8.jpg" alt=""></td>
-                                                    <td>Rhona Davidson</td>
-                                                    <td>Integration</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/10/14</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic9.jpg" alt=""></td>
-                                                    <td>Colleen Hurst</td>
-                                                    <td>Javascript Developer</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2009/09/15</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic10.jpg" alt=""></td>
-                                                    <td>Sonya Frost</td>
-                                                    <td>Software Engineer</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2008/12/13</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic1.jpg" alt=""></td>
-                                                    <td>Jena Gaines</td>
-                                                    <td>Office Manager</td>
-                                                    <td>Female</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2008/12/19</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic2.jpg" alt=""></td>
-                                                    <td>Quinn Flynn</td>
-                                                    <td>Support Lead</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2013/03/03</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic3.jpg" alt=""></td>
-                                                    <td>Charde Marshall</td>
-                                                    <td>Regional Director</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2008/10/16</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic4.jpg" alt=""></td>
-                                                    <td>Haley Kennedy</td>
-                                                    <td>Senior Marketing</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/12/18</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic5.jpg" alt=""></td>
-                                                    <td>Tatyana Fitzpatrick</td>
-                                                    <td>Regional Director</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/03/17</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic6.jpg" alt=""></td>
-                                                    <td>Michael Silva</td>
-                                                    <td>Marketing Designer</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/11/27</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic7.jpg" alt=""></td>
-                                                    <td>Paul Byrd</td>
-                                                    <td>Financial Officer</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/06/09</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic8.jpg" alt=""></td>
-                                                    <td>Gloria Little</td>
-                                                    <td>Systems Administrator</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2009/04/10</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic9.jpg" alt=""></td>
-                                                    <td>Bradley Greer</td>
-                                                    <td>Software Engineer</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/10/13</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic10.jpg" alt=""></td>
-                                                    <td>Dai Rios</td>
-                                                    <td>Personnel Lead</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2012/09/26</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic1.jpg" alt=""></td>
-                                                    <td>Jenette Caldwell</td>
-                                                    <td>Development Lead</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2011/09/03</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic2.jpg" alt=""></td>
-                                                    <td>Yuri Berry</td>
-                                                    <td>Marketing Officer</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2009/06/25</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic3.jpg" alt=""></td>
-                                                    <td>Caesar Vance</td>
-                                                    <td>Pre-Sales Support</td>
-                                                    <td>Male</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2011/12/12</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic4.jpg" alt=""></td>
-                                                    <td>Doris Wilder</td>
-                                                    <td>Sales Assistant</td>
-                                                    <td>Female</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/09/20</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic5.jpg" alt=""></td>
-                                                    <td>Angelica Ramos</td>
-                                                    <td>Executive Officer</td>
-                                                    <td>Male</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2009/10/09</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic6.jpg" alt=""></td>
-                                                    <td>Gavin Joyce</td>
-                                                    <td>Developer</td>
-                                                    <td>Female</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/12/22</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic7.jpg" alt=""></td>
-                                                    <td>Jennifer Chang</td>
-                                                    <td>Regional Director</td>
-                                                    <td>Male</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/11/14</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic8.jpg" alt=""></td>
-                                                    <td>Brenden Wagner</td>
-                                                    <td>Software Engineer</td>
-                                                    <td>Female</td>
-                                                    <td>B.TACH, M.TACH</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2011/06/07</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic9.jpg" alt=""></td>
-                                                    <td>Fiona Green</td>
-                                                    <td>Operating Officer</td>
-                                                    <td>Male</td>
-                                                    <td>B.A, B.C.A</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2010/03/11</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
-                                                <tr>
-                                                    <td><img class="rounded-circle" width="35"
-                                                            src="images/profile/small/pic10.jpg" alt=""></td>
-                                                    <td>Shou Itou</td>
-                                                    <td>Regional Marketing</td>
-                                                    <td>Female</td>
-                                                    <td>B.COM., M.COM.</td>
-                                                    <td><a href="javascript:void(0);"><strong>123 456 7890</strong></a></td>
-                                                    <td><a href="javascript:void(0);"><strong>info@example.com</strong></a>
-                                                    </td>
-                                                    <td>2011/08/14</td>
-                                                    <td>
-                                                        <div class="d-flex">
-                                                            <a href="#" class="btn btn-primary shadow btn-xs sharp me-1"><i
-                                                                    class="fas fa-pencil-alt"></i></a>
-                                                            <a href="#" class="btn btn-danger shadow btn-xs sharp"><i
-                                                                    class="fa fa-trash"></i></a>
-                                                        </div>
-                                                    </td>
-                                                </tr>
+                                            <tbody id="list-body">
+
                                             </tbody>
                                         </table>
                                     </div>
@@ -647,4 +49,65 @@
             </div>
             @include('admin.layouts.bottom-footer')
         </div>
+    @endsection
+    @section('section-script')
+        <script>
+            function loadlist() {
+                var html = "";
+                $.ajax({
+                    type: "GET",
+                    url: "{{ route('admin.listiing.show') }}",
+                    dataType: "JSON",
+                    success: function(response) {
+                        $.each(response, function(index, value) {
+                            html += '<tr>' +
+                                '<td>' + value['id'] + '</td>' +
+                                '<td>' + value['title'] + '</td>' +
+                                '<td>' + value['type'] + '</td>' +
+                                '<td>' + value['category'] + '</td>' +
+                                '<td>' + value['address'] + '</td>' +
+                                '<td>' + value['city'] + '</td>' +
+                                '<td>' + value['package'] + '</td>' +
+                                '<td>' + value['description'] + '</td>' +
+                                '<td>' + value['slug'] + '</td>' +
+                                '<td>' +
+                                '<div class="dropdown"><a href="#" class="btn btn-primary light sharp" data-bs-toggle="dropdown" aria-expanded="false"><svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="18px" height="18px" viewBox="0 0 24 24" version="1.1"><g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd"><rect x="0" y="0" width="24" height="24"></rect><circle fill="#000000" cx="5" cy="12" r="2"></circle><circle fill="#000000" cx="12" cy="12" r="2"> </circle><circle fill="#000000" cx="19" cy="12" r="2"> </circle></g></svg></a><ul class="dropdown-menu dropdown-menu-end" style="margin: 0px;"><li class="dropdown-item edit-cat" data-bs-toggle="modal" data-bs-target="#sendMessageModal" data-id="1"><i class="fa fa-plus text-primary me-2"></i>Edit</li><li class="dropdown-item del-cat" data-id="' +
+                                value['id'] +
+                                '" data-p-id="0"><i class="fa fa-ban text-primary me-2"></i>Delete</li></ul></div>' +
+                                ' </td>' + ' </tr>';
+                        });
+                        $("#list-body").html(html);
+                    }
+                });
+            }
+            loadlist();
+            $(document).on('click', '.del-cat', function(e) {
+                e.preventDefault();
+                var id = $(this).attr('data-id');
+                swal({
+                    title: "Are you sure?",
+                    text: "Once deleted, you will not be able to recover this Data!",
+                    icon: "warning",
+                    buttons: true,
+                    dangerMode: true,
+                }).then((willDelete) => {
+                    if (willDelete) {
+                        $.ajax({
+                            url: "{{ url('/admin/listing/delete') }}" + '/' + id,
+                            type: "GET",
+                            success: function(data) {
+                                swal("Poof! Your Data has been deleted!", {
+                                    icon: "success",
+                                });
+                                loadlist();
+                            },
+                            error: function(data) {
+                                swal("Oops", "We couldn't connect to the server!",
+                                    "error");
+                            }
+                        });
+                    }
+                });
+            });
+        </script>
     @endsection
